requirements.txt
Flask>=2.0.0
python-dotenv==1.1.0
redis==6.1.0
yt-dlp==2024.12.13
gunicorn==23.0.0
psycopg2-binary==2.9.9
elasticsearch==8.12.0
minio==7.2.5
Werkzeug==3.1.1
Jinja2==3.1.4
celery==5.3.4
APScheduler==3.10.4